backend/requirements.txt
colorama==0.4.4
grpcio==1.39.0
grpcio-tools==1.39.0
protobuf==3.17.3
six==1.16.0
uvloop==0.15.3
